Conditionally generate an extra javac action for kythe indexing
This additional action should not be depended upon by anything and only exists
so that kythe invocations can extract indexing metadata using the original
sources as input.

This symbol isn't guaranteed to succeed and builds must run with
`GROK_ELLIPSIS_BUILD` defined in addition to the per-library configuration.

The intended use-case for this is for source code that is exclusively pulled
into a `j2cl_library` and thus lacks any indexing. If the code is already pulled
into a `java_library` then this option should not be used.

PiperOrigin-RevId: 370964691