Fix IfStatement.clone().
If statement was not cloning its condition on clone() leading to an edge case in rewriting FunctionExpression; resulting in a "bad variable name" error in the wasm compilation due to referencing a variable that was not declared.

All the cloning logic needs to be audited and tested with more coverage.

PiperOrigin-RevId: 371820418