Fix missing erasure cast when unboxing is involved.
The pass that inserts the erasure related casts was completely ignoring all casts, since in general there is no need to add an erasure cast if the result will be explicitly cast to some other type. But the exception is that cast to primitive types are a different type of cast, since they might imply many things, like a type conversion or unboxing. In these cases an erasure cast might be needed see for example:

   List<Integer> l = ...;
   int i = (int) l.get(0);

Since l is actually a list of Object at runtime (due to erasure), a cast needs to be inserted before the unboxing operation is performed, resulting in:

   int i = (int) (Integer) l.get(0);

Implicit conversions like

   int i = l.get(0);

were already correctly handled.

PiperOrigin-RevId: 373232010