[WASM] Hande binary name used for entry point when it matches both static and non-static methods.
Instead of failing if the name matches a non-static method, do not match non-static methods. This will still on correctly reporting an error if there is a single method with the binary name and that method is not static.

There is also the overload problem where multiple methods with same name might be exported. In that case we will still exporting the first match, the CL is not addressing that.

PiperOrigin-RevId: 378030163