[WASM] Remove unnecessary newlines from the output.
Now, in StatementTranspiler each statement handler explicitly emits a newline before the statement, if it emits code for it.

In ExpressionTranspiler, expressions that prefer to be emitted in multiple lines (like MultiExpression and VariableDeclarationExpression) can emit newlines but should only do so after emitting some code at the start, otherwise there might be an empty line).

PiperOrigin-RevId: 383686602