Switch from Minion to simple shell diff test.
Current minion test is missing proper directory diffing and will require adding more complexity to fix that.
By switching simple shell diff we will be missing future opportunity to generate golden files from test results. On the other hand, the main reason for the golden tests are to ensure developers doesn't accidentally submit CLs with stale readables. We expect them to run `j2 gen` as part of their development process and check all the diffs before sending the CL, not after the fact. Hence missing sponge based update of goldens is not much of a loss.

PiperOrigin-RevId: 405758682