Handle postfix expressions according to JLS numeric promotion rules.
Postfix expressions are never a numeric promotion context according to JLS 5.6.1 and that was incorrectly handled in the ConversionContextVisitor. In practice this change is a noop, since postfix expressions that require other handling are always expanded.

There seems to be a gap in the ConversionContextVisitor regarding conversion in compound assignments, both unary and binary; but in practice all the compound assignments that need conversions are expanded before applying them in the current pipeline.

PiperOrigin-RevId: 410077475