Fix handling of unused AutoValue types.
Existing mechanism to preserve unused fields were causing some of the unused AutoValue types to be kept since prototype was kept alive until the late stages of JsCompiler optimizations. This change preserves the fields inside the constructor so that when the constructor is no longer used, the type is completely optimized away.

PiperOrigin-RevId: 410333738