Introduce MemberReference as a super class of FieldAccess, MethodCall and NewInstance with a common generic Builder.
This removes some redundant code in the AST nodes as well as in a couple of passes. In the past we have kept all the actual properties on the nodes on the concrete leaf classes leading to some duplication.

The drawback of moving some properties up the hierarchy is that it makes the implicit order in which the fields are visited opaque. One invariant that needs to be preserved in our passes is that nodes should be visited in evaluation order (i.e. in this case, visit the qualifier before the arguments). In this case since MemberReference has the qualifier and Invocation, its subclass, has the arguments the order is preserved; but it is no longer immediately obvious.

PiperOrigin-RevId: 410506652