Fix InternalCompilerError due to incorrect handling of a corner case involving type variables in the normalization of ForEachStatement.
The case in question is something like

    I iterable = ....
    for (E element : iterable) {
        ...
    }

where I is a type variable, possibly I extends Iterable<T>.

The normalization pass will insert a cast on iterable resulting in code like:
    I iterable = ....
    for (E element : (Iterable<E>) iterable) {
        ...
    }

The cast insertion infers the cast type (in this case Iterable<E>) from the loop variable type. The code incorrectly assumed that whatever type was present, it could call toBoxedType() to handle the case where the loop variable is a primitive, but that method is not implemented in TypeVariable triggering an InternalCompilerError.

Fixes https://github.com/google/j2cl/issues/140

PiperOrigin-RevId: 423388584