Fix the traversal order in DoWhileStatement.
AST nodes are meant to be traversed in lexical ordering (i.e. the order in which the code is written), and the order is provided by the ordering of the @Visitable field declarations in the class, which was incorrect for DoWhileStatement.

In the current state this change is effectively a noop.

PiperOrigin-RevId: 425998100