[J2KT] Split J2KT JRE into JRE emulation and JRE support targets
The javaemul annotations library needs the support code (the
transpiler will insert references to that code) but not the full JRE.
The JRE needs the annotations. Splitting the support code off of the
JRE avoids the dependency cycle.

It also lets us eliminate the weird “JRE for JVM” target which only
contains the support code (no JRE emulation necessary for JVM).

PiperOrigin-RevId: 437840118