[J2KT] Add a placeholder `java.lang.Class` for J2KT Native
This is just enough to make `integration/testing/Asserts*` compile on J2KT Native.
Without the asserts, we can’t run any integration tests.

This placeholder doesn’t quite match the existing filename patterns.
To benefit from the implicit import for j2kt extension methods, it has
to be in `javaemul.lang`. But it can’t be included in the support library
on JVM because it’s incompatible with the real JRE (or kotlin.jvm.*)
implementation. So I’m including it in the J2KT Native JRE emulation library.

PiperOrigin-RevId: 438075999