Optimize Array initialization.
The new implementation is 4-5x faster than earlier implementation for int[].

Normally, I would expect an array that's only composed of ints to be more optimal when they are added one at a time vs. upfront length since in the former case it seems easier for V8 to keep the more optimal/compact array representation.

I can think of 2 reasons why the benchmarks doesn't reflect that:
1. We were not getting benefit from the compact arrays due to expandos.
2. Array(n) is somehow still magically optimized by V8

PiperOrigin-RevId: 441599914