[WASM] Emit immutable fields in structs.
Both binaryen and v8 benefit from knowning if fields are immutable. This cl add a global analysis to determine which fields are effectively immutable and moves their initialization to the struct.new instantiation.

The optimizations is done by tracking how the initialization values flow through constructor delegation chains and hoisting the field assignment to each constructor factory in the class and all subclasses.

Fields that can potentially be observed uninitialized are removed from consideration.

PiperOrigin-RevId: 447709653