Suggested tests to add to J2CL, prompted by a bug in GWT
A patch made to GWT last year introduced a bug in turning Longs to
strings - while the same patch was made to J2CL, it did not have an
issue with signs, so these tests should pass as-is. With that said,
there was no test coverage for values that are positive as longs, but
negative when treated as an int, with all high bits set.