Add more edge cases to long testing.
A patch made to GWT last year introduced a bug in turning Longs to
strings - while the same patch was made to J2CL, it did not have an
issue with signs, so these tests should pass as-is. With that said,
there was no test coverage for values that are positive as longs, but
negative when treated as an int, with all high bits set.

See https://github.com/gwtproject/gwt/pull/9769

Closes https://github.com/google/j2cl/pull/155

PiperOrigin-RevId: 464563372