Provide access to setup/teardown methods via TestCaseViolator instead of non-standard APIs.
Earlier, to run junit3 tests, we had hidden setUp and tearDown method to expose the protected setUp and tearDown instance methods.
However j2kt junit tests need to rely on official java junit and those APIs are not available.
This CL follows an alternative design where a TestCaseViolater added in the same package as junit framework so we can call these methods outside.

PiperOrigin-RevId: 480724581