[WASM] Optimize equalsIgnoreCase to avoid repetitive back and forths to JS for heavy non-ascii scenario.
This makes equalsIgnoreCase 6x faster for non-ascii benchmarks and 2x for ascii. The later is very likely due to better inlining of the new implementation and it is now faster than JVM.

Being said that the benefits are not clear for the average case since some non-worst case scenarios may perform worse due to eager toLowerCase for the whole string.

PiperOrigin-RevId: 484414444