Rollforward of Fix nullability inconsistencies in supertypes and method descriptors.
Supertype type arguments need to be created using the @NullMarked context of the type declaration.

Method descriptors where created inconsistently, so for consistence use the nullmarking of the enclosing class declaration.

This does not solve the major issue of inferred type arguments, but fixes inconsistencies so that the pipeline can rely on facts like there is only one method descriptor that represents a method declaration.

PiperOrigin-RevId: 486014545