[WASM] Replace the generic array for itables by a struct and create specialized struct types for each itable type.
The itable is now a struct of "(ref null struct)" fields with as many fields as the maximum size of the former itable array. For each type this struct gets specialized so that for the slots that have an interface for that type the field will be types as "(ref $interface.vtable)" and these specialized itable also are declared hierarchically following the class hierarchy.

As a side effect there is no more array bounds checking in the interface dispatch or instanceof.

PiperOrigin-RevId: 490144928