Workaround a bug due to the issue of specializing type variable bounds.
MethodDescriptor.getOverriddenMethods() used getPolymorphicMethods() to retrieve the list of potential overridden methods. getPolymorphicMethods() would included some descriptors that are incorrectly specialized due to the type variable bound specialization bug.

The effect of that bug is that in some corner cases it would miss overridden methods.

The change here just adds all the declared methods to the list of potential overrides. Such methods are constructed in the frontend, and in the case of JDT, don't use the specializeTypeVariables so they include the correct specialization.

PiperOrigin-RevId: 490404307