Use the new Binaryen flag for "closed world".
This flag does nothing atm, but Binaryen will switch to the safe default of an open world. At that point the flag will be needed to keep optimizing the way we have been so far. So adding the flag now will do nothing now, and make that later Binaryen change do nothing as well, and avoid a regression.

PiperOrigin-RevId: 491703972