[KT] Add a rewriting pass to return kotlin.Nothing at callsites
Kotlin treats calls to functions that return `Nothing` as if it were a throw
(since those functions will never return). These semantics don't hold up for
Java or JS where you must have a return or explicitly throw.

I paired this with another pass that will remove unreachable statements in a
block by trimming off statements beyond a return, throw, or nothing statement.
This is required as the JSC can get upset if you early return but then have
further statements as some types can be in a temporal dead zone, particularly in
a constructor.

To workaround this we can simply return the result of the call, which will
ultimately be `undefined` in JS.

PiperOrigin-RevId: 503565131