Small change to MemberDescriptor.isExtern().
An extern member descriptor is a member has an extern namespace. The restriction checker enforces the restriction that extern members need to be declared native so the native check in isExtern was unnecessary.

Moreover the kotlin frontend does not keep track of whether a Java method referenced from kotlin is declared as native.

PiperOrigin-RevId: 509856635