[WASM] Allow `j2wasm_library` to accept `js_library` deps.
This change allows `js_library` deps to be included in `j2wasm_library`, but has no other visible effects until JS imports generation is implemented.

The ultimate goal is to allow JS imports generation by allowing the transpiler to generate a `js_library`, containing the import mapping, depending on the provided `js_library`s and referencing contained JS types.

For this, I have `j2wasm_library` return a provider exporting all provided `js_library` deps (It has to be "exports" to avoid breaking strict deps. This can later be "deps" instead of "exports" upon modular compilation if the js imports are generated at the -library level). These exports are currently consumed by `j2wasm_application` for it to later use as a dep for a JS provider which contains the JS import mapping.

PiperOrigin-RevId: 513886932