[WASM] Optimize String.compareToIgnoreCase.
This loosely based on similar optimization that I have implemented for String.equalsIgnoreCase where we optimistically do the comparison in Wasm until we see a non-ascii character.

This applies similar optimization to both stringref and non-stringref versions with the exception that non-stringref version fully delegates to JS when it reaches the non-ascii character since there is no point of copying the string back. Again here it follows the same approach in String.equalsIgnoreCase.

PiperOrigin-RevId: 516609752