[JS] Correctly stamp the array type on bridges for jsmethods with varargs.
The semantics of a jsmethod with varargs allows it to be called with an unstamped array that might not match the varargs declarations. For these cases, arrays received as varargs by JsMethods are first stamped to the declared type.

The current code does not take into account that some of those might come from bridges that have a more general array type and would throw class cast exception when forwarding to the bridged method.

PiperOrigin-RevId: 517497996