Remove unnecessary condition in MethodDescriptor.isOverride.
MethodDescriptor.isOverride was explicilty returning false for methods that reduce the visibility. Code with overrides that reduce the visibility is already rejected by Java, hence there is no need to check.

Additionally, in some cases in the boundary between Java and Kotlin that situation might happen and the method that reduces the visibility still needs to be considered an override (see Kotlin's AbstractCollection#toArray that is protected and J2CLs JRE Collection#toArray that is public)

PiperOrigin-RevId: 522227899