Don't use isEffectivelyFinal for variables in the front end, instead run MakeVariablesFinal early on in the pipeline.
JDT will mark variables that are written to as effectively final if the code that writes to them is unreachable leading to an inconsistent AST affecting j2kt.

Since it is dead code it would seem that this is not very relevant but there are patterns like

```
if (DEBUG) {
  for (int i=0;...;i++) ...
}
```

where dead code does occur.

PiperOrigin-RevId: 523521746