Replace invalid UTF-8 characters rather than throw in String constructors
The contract of the overloads that accept a `Charset` (the literal, not a string
of the charset name) is explicit in that invalid encodings should be replaced
with the Charset's default replacement char [1]. The existing implementation
does not do this and throws an IAE instead.

All other overloads are unspecified in the behavior, but other JRE
implementations opt to replace rather than throw.

While I improving test coverage I also realized that we weren't handling
overlong (i.e. zero-padded) encodings correctly. While the UTF-8 spec is silent
on the behavior for decoding, encoders are required to emit the shortest
possible encoding. Due to this decoder implementations treat overlong encodings
as invalid.

[1]:
https://docs.oracle.com/javase/8/docs/api/java/lang/String.html#String-byte:A-int-int-java.nio.charset.Charset-

PiperOrigin-RevId: 524344606