Correct logic for determining overlong encodings
The previous logic was wrong as it only looks at the bits from the highest byte.
However, we actually want to look at the highest octet, which may span multiple
bytes. For example, U+0800 would have incorrectly tripped this check.

The easiest way to handle this is to decode the codepoint and then just perform
a check against the well-known codepoints that fit within 1, 2, or 3 bytes.

PiperOrigin-RevId: 525436201