Fix UniqueNamesResolver to emit unique names for type parameters of a method.
UniqueNamesResolver was relying on the fact that a type variable appears only once as a declaration either as the type parameters of a class or of a method. However that invariant is not kept when methods are devirtualized, where the type variables declared in the enclosing class are also type parameters of the devirtualized method.

UniqueNamesResolver assigns a unique name to each type variable and keeps track if they appear in unrelated contexts and can be assigned the same name. Before this patch, the first time it encountered the declaration was assigning the name and registering it to the current context.

This patch now correctly registers the type variable with each context instead of just the first one that introduced it.

PiperOrigin-RevId: 526042932