Ensure we don't exclude targets being exported if they're also deps
We're attempting to manually recreate the classpath we'll need to compile the
given target. Unfortunately targets included in exports have their compile_jars
included in the compile_jars for the targeting being compiled.

This is problem if we have a target in both exports **and** deps as we'll
exclude that compile_jar.

PiperOrigin-RevId: 527953296