Change in bridge method computation to not create bridges that target other bridges.
The algorithm already determines what is the right implementation from source, there is no need to target a specialized bridge instead of the actual implementation in the source.

This results in the creation of less number of bridge methods and some bridges might now target a further away implementation. This also means that there are less erasure casts and that the behavior on bridges is more similar to javac.

PiperOrigin-RevId: 528888432