Add an option to preserve type parameters when converting method references to lambdas.
Functional interfaces with type parameters in the functional method can not be instantiated directly with lambdas, but can be instantiated with method references (or classes). There is a pass that converts method references into lambdas, and in most backends needs to remove the type parameters; however j2kt will use the lambda to synthesize an anonymous class and needs to know the method type parameters.

PiperOrigin-RevId: 535492269