[WASM] Generate separate imports for different signatures for js methods.
I had to change the merge logic (remove it) and generate a separate import for each signature. Example: If we have both "Date.constructor(number)" and "Date.constructor(number, number)" this results in 2 imports "Date.constructor$1" and "Date.constructor$2". These 2 calls are not equivalent.

The logic to determine the index is as follows:
- If the method can be imported using a method reference or is a js property, do not add an index.
- Otherwise, use the size of the parameter list. A compatibility check is still performed for methods with the same number of parameters. We currently do not have any examples of incompatible methods with the same number of parameters (hence, this cl has it report an error if encountered).

PiperOrigin-RevId: 537328793