[KT] Optimize Kotlin array constructor by threading the initializer in NewArray
This modifies the J2CL AST to be aware of an initializer lambda (or method
reference) that can be used to initialize the values of the array when it's
created.

In turn this is threaded into the `vmbootstrap.Arrays.$create()` call where it
will short-circuit recursive allocation in favor of executing the initializer
instead.

PiperOrigin-RevId: 538291909