[WASM] Handle unqualified `super` in lambdas correctly.
When the lambda implementor class is introduced, unqualified super method calls (that target the super type of the class enclosing the lambda) need to be marked as static dispatch; because the transformation introduces a new enclosing class, these super method calls, if not fixed, would target a super method of the lambda implementor (due the inconsistent modeling, it ends up doing a static dispatch but with the incorrect receiver).

The modeling of super method calls is quite awkward. Super method calls that target super classes of outer classes are normalized to have static dispatch and the qualifier is changed to `ThisReference`. This is done in `NormalizeSuperMemberReferences` very early on and after that every super that targets an outer super needs to follow this form.

PiperOrigin-RevId: 538469163