Cleanup names of passes.
Pass names should read like a action to be performed, e.g. `OptimizeEnums`, `NormalizeCasts`, `ImplementArrayAccesses`.

To mitigate the ambiguity of having both a Kotlin frontend and a Kotlin backend, we will use the J2kt suffix for passes that are specific to the Kotlin backend, i.e. passes that are j2kt versions of existing passes (like `InsertWideningConversionsJ2kt`), or passes that are not potentially usable as general transformations (like `PropagateNullabilityJ2kt`).

The same applies for Wasm specific passes.

PiperOrigin-RevId: 540135429