[KT] Disallow use of the `@JsProperty`/`@JsIgnore` on Kotlin fields.
This is a weird edge case of Java/Kotlin annotation interop. Since there
annotations target both `FIELD` and `METHOD`, it's somewhat ambiguous how this
would apply to a Kotlin property. In practice it ends up applying to the backing
field of the property. However, the backing field is not part of the Kotlin ABI,
so downstream targets will never know the corresponding property had annotations
on it.

The workaround is to annotate the getter/setter directly. This can be done by
either annotating it directly (if you defined a getter):

```kotlin
val foo: String
  @JsProperty get() = "bar"
```

Or by specifying the use-site target:

```kotlin
@get:JsProperty val foo: String
  get() = "bar"
```

PiperOrigin-RevId: 542034493