Stop passing leafTypeInitialValue explicitly on Array creation.
Instead leafTypeInitialValue will be available on the constructor where applicable (i.e. primitives).

This is slightly more efficient but also in the future will help us to make ArrayHelper.clone safe (which is currently not if the new size is larger).

Since I was here, I fixed all the linter errors related to missing explicit nullability on types.

PiperOrigin-RevId: 547048030