Support finding `.native.js` files based on fully qualified names
We currently rely upon the `native.js` files being adjacent to each based on
relative paths. With this change we'll support matching on a symbols fully
qualified name if it's contained within the filename, regardless of the rest of
the path.

For example, the file `com/google/foo/bar/Buzz.java` with the package statement
`package com.google.foo.bar` containing the class `class Buzz` can have a
`native.js` in any path so long as the filename is
`com.google.foo.bar.Buzz.native.js`.

Note: the fully qualified name is always the Java fully qualified name.
The JS name does not apply.
PiperOrigin-RevId: 554896423