Only match fully qualified native.js files if there's >= 1 package segment
If there's no package segment then it's ambiguous if a file is intended to be
match based on relative path or fully qualified path. This can be problematic
if there's multiple `native.js` files with the same name (but different
packages) that were intending to match on relative path.

For example, `foo/bar/Platform.native.js` and `baz/qux/Platform.native.js`
would both be considered the `native.js` file for the fully qualified class
`Platform`.

PiperOrigin-RevId: 555752990