Rewrite EnumSet/EnumMap based on TreeMap.
We are slightly out-of-contract here as iteration is no longer weakly consistent; instead we fail-fast. An alternative would be using large Array or ArrayList where ordinal is index however that has significant memory cost especially for large enums (which is not rare in protobufs).

This removes dependency on JavaScript resizing behavior in these classes.

PiperOrigin-RevId: 555934812