[WASM] Stop relying on JavaScript resizing behavior in BitSet.
Since we need to start tracking some kind of logical length, I made sure it is the actual last "set" words and updated the algorithms to take advantage of that.

Behind the scenes we are still relying on resizing but it is now hidden behind ArrayHelper that I will replace later on.

PiperOrigin-RevId: 556050093