Fix j2wasm_application defaults to have check-level set to NORMAL.
Also fixes integration tests to have the same defaults and correctly pipe the defines.

We are currently not following the best practice of setting same check-level for both dev and optimized but this is due to having a different default for optimized. We should probably have MINIMAL as the default for both JS and WASM and get it overridden at app&test level. This is something we should address at some point but not within the scope of this CL.

PiperOrigin-RevId: 559196459