Render native field of native `JsType` as `lateinit`.
Field on native JsType are left uninitialized as they are defined in Javascript.
Kotlin enforces field initialization during their declaration. To workaround this issue, we will mark uninitialized native field as `lateinit`. J2CL does not produce any code for native `JsType` and this has no effect on the final generated code.

A better option would be to use a dummy function named `definesExternally()` that can be used as default initializer.

PiperOrigin-RevId: 571162134