Avoid promoting implicit default constructor to public.
After j2kt transpilation, all types are conservatively translated in Kotlin as public. That directly impacts the implicit constructor that is also promoted to public.
When the type is JsType, that changes the Javascript contract of the type as the implicit default constructor becomes a `JsConstructor`.

This cl will render a default primary constructor with a more restrictive visibility when the class has no constructor and is not public. The logic to decide om the visibility of the constructor is the same than the one used for class members.

PiperOrigin-RevId: 576538147