[WASM] In preparation for non-native JsEnum optimizations in Wasm, introduce a new InternalPreconditions.checkNotNull(boolean) and use it for the JsEnum null checks (e.g., in NormalizeSwitchStatements).
This is needed because we need to be able to normalize null checks for JsEnums separately.

For example: "checkNotNull(jsenum)" does not work because we want to avoid boxing the jsenum. We must emit "checkNotNull(jsenum != null)" and then Wasm can rewrite it into something like "checkNotNull(jsenum != MIN_VALUE)"

PiperOrigin-RevId: 576943966