Start emitting references to `kotlin.js.definedExternally` as initializers of uninitialized native fields.
Right now we consider uninitialized native fields as lateinit. That brings some issue as KoltinC add getter and setter for the property. Those getter/setter functions are not part of the exposed Javascript API and had to be marked as JsOverlay. Also they can conflict with native getter/setter methods exposed by the native JsType.
`kotlin.js.definedExternally` has been added recently to the J2CL stdlib and its purpose is to be used to dummy initialize native fields of native JsType.

PiperOrigin-RevId: 577302147