Enable j2kt-web on `j2cl_library` that have no sources.
A `j2cl_library` without srcs may export other `j2cl_library`. If we do not enable j2kt-web for those libraries we lack the J2kt provider that reexport the exported library.
When we use those libraries as dependencies, we end up creating a J2kt provider based on the J2CL provider and using the wrong Java provider for the Java compilation made before the J2kt transpilation.

PiperOrigin-RevId: 579697564