[WASM] Separates the responsibilities for producing outputs and synthesizing .wat code.
Formerly `WasmOutputGenerator` would sinthezise the top level .wat code and produce all the files in a wasm transpilation. This cl renames `WasmOutputGenerator` into `WasmOutputGenerationStage` whose responsibility is only to produce the output files and control how the output is produced; and introduces `WasmConstructsGenerator` whose responsibility is to produce .wat constructs for AST and runtime needs.

`WasmOutputGenerationStage` is also responsible for the use and lifetime of the SourceBuilder objects.

PiperOrigin-RevId: 583516872