Automated g4 rollback of commit 4cf2508c0c678554d3727bf7edd940079aa5917e.
*** Reason for rollback ***

This is not needed anymore if we do not enable j2kt-web readables. In addition to that, native js files in `nativeinjection` readable should not be renamed as we lost test coverage.

*** Original change description ***

Rename native files to make it compliant with j2kt-web.

We will be turning j2kt-web readables on by default soon. The Kotlin frontend requires that the native files are either located in directory reflecting the package statement of the native JsType or that the name of the file is the FQN of the native JsType.

***

PiperOrigin-RevId: 586381761