[WASM] Use factory methods when implementing arrays as classes.
The modular backend only constructs a full structural view of the classes that are declared in the library being compiled (which is needed to emit NewInstances). Classes are normalized to hide NewInstances behind factory methods but the pass that implements arrays as classes runs afterwars and was introducing a NewInstance in the usage side.

PiperOrigin-RevId: 587076412