[WASM] Small fix in wasm layout constructions that only affects modular compilation.
The issue would surface in a scenario where the layot for a type `A` that is not in the currently library was being created and that had a supertype `B` also not in the current library that was seen for the first time. In that scenario the wasm layout for type `A` would have `null` as the layout for its supertype.

PiperOrigin-RevId: 588924936