[WASM] Avoid triggering clinit from internal calls in class
This handles basic inner cycles for clinit due to calls to methods.

Consider following:

```
class Foo {
   // initialized in Foo.clinit
   public static final Foo instance = new Foo();

   // Also calls Foo.clinit since it is public
   public Foo() {}
}
```

We essentially have two entry points in the classes that we need to clinit. This introduces a cycle in clinit (clinit call ctor, ctor calls clinit) and we cannot hoist instance to global because it is no longer trivial.

This could also happen with following pattern:

```
class Foo {
   // initialized in Foo.clinit
   public static final int instance = someHelper();

   // Also calls Foo.clinit since it is public
   public static int someHelper() { ... }
}
```

To overcome this problem, when there is a public method that triggers clinit, we can make it private and add a public one that calls clinit. Then re-writing Internal calls go through private one so they don't trigger clinit. Since this only requires local knowledge, it is compatible with modular compilation.

PiperOrigin-RevId: 590882728