Support pre-super() call statements in a JsConstructor and delegates
This mostly just means inlining any of these statements into the synthesized ES6
ctor before the `super()` call.

For non-primary ctors, the statements will be moved into the synthesized factory
method in the same fashion.

The only quirk I encountered was with how we resolve captures for inner classes.
We were putting these before the super call, even though it wouldn't actually
execute in that order for `JsConstructor`s. I reworked this to make it a bit more
explicit and not rely upon the `super()` call just being plucked out.

PiperOrigin-RevId: 592060804