Disable logging by default for optimized builds.
We already disabled the simple log handler so if the app doesn't set any custom handler, then the logging wouldn't work anyway (in that case they can also configure the log level to enable it).

When I was there, I also updated integration tests to have the same default as JavaScript.

PiperOrigin-RevId: 592116664