Remove invalid reference to `this` as an enclosing instance for anonymous classes  in parameters to super/this construtor calls.
Anonymous classes are considered to capture their enclosing instance except when declared in a static method. In constructors those classes capture the enclosing instance unless they appear as a parameter to a super/this constructor call. Since the type model object is constructed independent from the code, anonymous classes that are declared in a parameter to a super/this constructor call are considered as capturing the enclosing instance.

This cl works around the problem by changing the way implicit qualifiers are computed, and in this cases the anonymous class will receive a null enclosing instance instead of a this reference. Such value will never be referenced since both Java and Kotlin forbid referencing `this` impliclitly or explicitly in these situation.

PiperOrigin-RevId: 600956219