[J2KT] Create forwarding methods for j.l.Object methods declared by interfaces.
Kotlin doesn't automatically inherit such methods from "Any" while implementing interfaces. Existing solution was removing them from interfaces but that breaks the contract with JavaScript but also not a lossless conversion.

Currently this is breaking native compilations so I guarded it with a flag and restricted it to only JsType to unblock current work.

PiperOrigin-RevId: 602923374