[WASM] Make JS errors catchable in Wasm.
To do that, this changes the exception tag used for exceptions to JSTag and start also throwing JS errors behind the scenes for Wasm exceptions. This model makes it closer to the modeling in J2CL/JS.

There are some cleanup opportunities in the way we model exceptions and make JS and Wasm runtimes closer but for now keeping this CL minimal to get something working.

PiperOrigin-RevId: 606387826