[WASM] Add a modular version of compile time constant propagation.
This pass replaces the `PropagateConstants` pass that is present in monolithic compilation, which relies on seeing the field initializations in the AST. In constrast the new pass only relies on the type model.

This pass is needed because of the way static fields are represented and initialized in J2Wasm.

Compile time constants replace the default value in static field initializations. In Wasm static field are modeled as globals and their initialization is performed in the global declaration. Global declarations only allow constant expressions as initializers, and string literals which end up being method calls are not allowed. Instead the values for these compile time constant fields are propagated to all the places where they are referred and the field is removed.

Additionally, this also allow to reuse the switch normalization passes which rely on synthetic compile time constants that represent the ordinal values.

PiperOrigin-RevId: 606417645