Simplify how privateInitError is handled.
This removes the side channel for exception constructors where we were intercepting privateInitError to use our passed in value. I don't know since when - probably after GWT split - but this is no longer necessary. We can just call privateInitError manually in Throwable.of to override the value. Only negative is we are setting Java exception on the dropped error object but that looks like a minor issue. A bigger problem there is the call to privateInitError with a thrown away Error object (generated by compiler), which is already the case.

A better solution to that problem would be preventing privateInitError to be called at the first place (maybe via custom subclasses) but that is out of scope of my intention with this CL.

PiperOrigin-RevId: 606710442