Process postfix not-null assertions for context type conversion
This revealed that the `InsertNotNullAssertions` pass would start emitting
duplicate not-null postfix expressions since it blasts them out to various
places first, and _then_ adds them based on context type conversion.
Unfortunately when that happens they will not know that they're already in
a non-null assertion as they only see the operand.

We can avoid this issue by swapping the order of the passes so that we put them
in type conversion contexts first, and then walk the AST to place them in
specific locations. When the latter occurs we can skip if it's already a
not-null assertion.

PiperOrigin-RevId: 609471059